CAMEL-10532 Extract ContractAdvice as an individual file
